pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lium-cli"
version = "0.1.0"
description = "A custom CLI tool for Lium. Lium CLI provides tools for interacting with the Lium ecosystem from the command line."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Lium", email = "support@lium.ai"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent"
]
dependencies = [
    "typer>=0.12,<0.16",
    "rich>=13.7,<15.0",
    "pycryptodome>=3.0.0,<4.0.0",
    "PyYAML~=6.0.1",
    "aiohttp~=3.10.2",
    "backoff~=2.2.1",
    "GitPython>=3.0.0",
    "async-substrate-interface>=1.1.0",
    "scalecodec==1.2.11",
    "plotly>=6.0.0",
    "plotille>=5.0.0",
    "fuzzywuzzy~=0.18.0",
    "python-Levenshtein",
    "netaddr~=1.3.0",
    "Jinja2",
    "pywry>=0.6.2",
    "async-property==0.2.2",
    "bittensor-cli>=8.4.3",
    "paramiko==3.5.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "wheel"]

[project.scripts]
lium = "lium_cli.cli:main"

[tool.hatch.metadata]
package-dir = {"lium_cli" = "lium_cli"}

[project.urls]
Homepage = "https://github.com/Datura-ai/lium-cli"
Repository = "https://github.com/Datura-ai/lium-cli"
Issues = "https://github.com/Datura-ai/lium-cli/issues"